Add all env. vars. used to pass mojo lib. search paths to tutorial
The env. var. shown in the tutorial section of the development document
for the Stdlib is (at least in some cases) insufficient to point `mojo`
to a newly build stdlib.mjpkg. This adds all the same variable that are
passed in the `lit` config to hopefully get all edge cases covered and
make getting stared easier.

Signed-off-by: Hristo I. Gueorguiev <53634432+izo0x90@users.noreply.github.com>