[stdlib] `List.extend` uses `origin` on its `UnsafePointer` logic.
Hello,
this small commit is to integrate the `origin` parameter into
the existing implementation.

(during logic on `UnsafePointer` to the `other` arg)

For example, if `other^.__del__()` is added before the loop,
an error comes on `move_pointee_into`:
"uninitialized value other"

(preventing moving of values out of an maybe freed pointer)

The `origin` parameter makes sure that `other` is not `__del__` while

we use the `UnsafePointer` to it.

Not sure if needed here, but it feels like a nice thing ! 👍

Signed-off-by: rd4com <144297616+rd4com@users.noreply.github.com>